requirements.txt
asgiref==3.3.0
certifi==2020.11.8
chardet==3.0.4
Django==3.1.3
django-stdimage==5.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
Pillow==8.0.1
progressbar2==3.53.1
psycopg2==2.8.6
python-dotenv==0.15.0
python-utils==2.4.0
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.1
whitenoise==5.2.0
